backend/requirements.txt
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
psycopg2==2.8.6
django_fsm==2.7.1
django_cors_headers==3.7.0
redis==3.5.3
celery==5.0.5
docker==4.4.4
pyyaml==5.4.1
django-model-utils==4.1.1